+++ selfdrive/loggerd/bootlog.cc
@@ -56,6 +56,7 @@ static kj::Array<capnp::word> build_boot_log() {
 }
 
 int main(int argc, char** argv) {
+  clear_locks(LOG_ROOT);
 
   const std::string path = LOG_ROOT + "/boot/" + logger_get_route_name() + ".bz2";
   LOGW("bootlog to %s", path.c_str());
