+++ selfdrive/loggerd/loggerd.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <ftw.h>
 #include <unistd.h>
 
 #include <atomic>
@@ -115,7 +114,6 @@ struct LoggerdState {
   bool camera_synced[WideRoadCam + 1] = {};
 };
 
-void clear_locks(const std::string log_root);
 bool sync_encoders(LoggerdState *s, CameraType cam_type, uint32_t frame_id);
 bool trigger_rotate_if_needed(LoggerdState *s, int cur_seg, uint32_t frame_id);
 void rotate_if_needed(LoggerdState *s);
